Updated and organized the import paths
