+++ frontend/src/components/common/skeletons/skeletonBox.tsx
@@ -1,4 +1,5 @@
 import React from 'react';
+
 import styled, { keyframes } from 'styled-components';
 
 interface SkeletonBoxProps {
